moving around the components
