+++ src/components/DybbuksAndDreidels/ActionDreidels.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Box, Button, Container, Grid, Tooltip, Typography } from '@mui/material';
 import HelpIcon from '@mui/icons-material/Help';
-import { Dreidel, Side } from '../components/Dreidel';
+import { Dreidel, Side } from './Dreidel';
 
 export const ActionDreidels = () => {
     const [actionTimer, setActionTimer] = useState<number>(80);
