+++ src/pages/DybbuksAndDreidels.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { Button, Grid, Typography } from '@mui/material';
-import { FAQ } from '../components/FAQ';
+import { FAQ } from '../components/DybbuksAndDreidels/FAQ';
 import {default as DybbuksAndDreidelsCover} from '../images/DybbuksAndDreidelsCoverFront.jpg';
